Create hard-reject.siv
Saw your post on Reddit. Neat idea.

This is my first Sieve filter. It's for senders that you don't want emailing you, and if they do email you, you don't even want to know about it.

It first checks if a sender is in the contact group REJECTS. If so, then it sends the sender a standard reject, which will delete the incoming message completely and send the sender the reject message from the same email address that the original incoming message was addressed to.

But then you have a message in your Sent messages with the subject "Auto: <original message subject>". Maybe you don't even want to know there was ever an email to begin with, so the "elsif" part watches for that outgoing reject messages, and removes it from your Sent messages.

Now, this script needs a bit of work (perhaps due to a but in ProtonMail??). That last "discard" doesn't delete the message permanently but puts it in your Trash. From reading the Sieve RFCs and ProtonMail docs, it seems like this shouldn't be occurring, i.e. it should be deleted completely. My workaround for this seeming bug is to add this 1 day expiration. So it should be deleted permanently after 1 day, but I just wrote this so I'll know tomorrow if this actually happens. I haven't tried with 0 days.

I've tested this with a few throwaway email addresses and works well enough. I think it's inefficient and has extraneous imports, but I'm not completely sure which imports to keep and which to remove. I'll play with it later maybe.

Anyway I think it's a decent first script for your repo.